+++ resources/views/livewire/resource/create.blade.php
@@ -1,11 +1,16 @@
 <div class="">
-    @section('title', $model->createHeaderTitle())
+    @section('title', e($model->createHeaderTitle()))
 
 
     @if(!$inModal)
+    @php($returnUrl = $model->createReturnUrl())
     <x-aura::breadcrumbs>
         <x-aura::breadcrumbs.li :href="route('aura.dashboard')" title="" icon="dashboard" iconClass="text-gray-500 w-6 h-6 mr-0" />
-        <x-aura::breadcrumbs.li :href="route($model->createReturnRoute())" :title="__($model->getPluralName())" />
+        @if ($returnUrl)
+            <x-aura::breadcrumbs.li :href="$returnUrl" :title="__($model->getPluralName())" />
+        @else
+            <x-aura::breadcrumbs.li :title="__($model->getPluralName())" />
+        @endif
         <x-aura::breadcrumbs.li :title="$model->createHeaderTitle()" />
     </x-aura::breadcrumbs>
     @endif